+++ README.md
@@ -10,7 +10,7 @@
   - [x] UpdateJdbcTemplate, InsertJdbcTemplate에 템플릿 메서드 패턴 적용
   - [x] 도메인 의존도 제거
   - [x] SelectJdbcTemplate 클래스 추출
-  - [ ] SelectJdbcTemplate에 템플릿 메서드 패턴 적용
+  - [x] SelectJdbcTemplate에 템플릿 메서드 패턴 적용
   - [ ] mapRow, setValues 인터페이스로 분리
   - [ ] Unchecked Exception 추가
   - [ ] RowMapper 제네릭 사용하도록 수정
