+++ app/src/main/java/com/techcourse/dao/JdbcTemplate.java
@@ -1,12 +1,12 @@
 package com.techcourse.dao;
 
-import com.techcourse.domain.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 
 public abstract class JdbcTemplate {
@@ -15,6 +15,8 @@ public abstract class JdbcTemplate {
     abstract String createQuery();
 
     abstract DataSource getDataSource();
+
+    abstract Object mapRow(ResultSet rs) throws SQLException;
 
     abstract void setValues(PreparedStatement pstmt) throws SQLException;
 
@@ -47,5 +49,48 @@ public abstract class JdbcTemplate {
                 }
             } catch (SQLException ignored) {}
         }
+    }
+
+    public Object query() {
+        final DataSource dataSource = getDataSource();
+        final String sql = createQuery();
+
+        Connection conn = null;
+        PreparedStatement pstmt = null;
+        ResultSet rs = null;
+        try {
+            conn = dataSource.getConnection();
+            pstmt = conn.prepareStatement(sql);
+            setValues(pstmt);
+            rs = executeQuery(pstmt);
+            log.debug("query : {}", sql);
+
+            return mapRow(rs);
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        } finally {
+            try {
+                if (rs != null) {
+                    rs.close();
+                }
+            } catch (SQLException ignored) {}
+
+            try {
+                if (pstmt != null) {
+                    pstmt.close();
+                }
+            } catch (SQLException ignored) {}
+
+            try {
+                if (conn != null) {
+                    conn.close();
+                }
+            } catch (SQLException ignored) {}
+        }
+    }
+
+    private ResultSet executeQuery(PreparedStatement pstmt) throws SQLException {
+        return pstmt.executeQuery();
     }
 }
